fix filepath None

+++ depyf/explain/patched_lazy_format_graph_code.py
@@ -26,7 +26,7 @@ def patched_lazy_format_graph_code(name, gm, maybe_id=None):
         commented_src += "".join(["# " + line +
                                  "\n" for line in src.splitlines()])
         src = simple_code + commented_src
-    if os.path.exists(filepath):
+    if filepath and os.path.exists(filepath):
         os.remove(filepath)
     new_filepath = write_code_to_file_template(
         src, os.path.dirname(filepath) + "/" + file_name + " " + "%s" + ".py")
